generalize multi key view args provider
